+++ packages/Doctrine/UnalterableMigrations/src/UnalterableMigrationInterface.php
@@ -1,16 +1,14 @@
 <?php declare(strict_types=1);
 
-namespace App;
-
 namespace PetrKnap\Doctrine\UnalterableMigrations;
 
-interface UnalterableSqlMigrationInterface
+interface UnalterableMigrationInterface
 {
     const REMOVE_PARENT = '-- remove parent';
 
     public function getParentClassName(): ?string;
 
-    public function getParent(): ?UnalterableSqlMigrationInterface;
+    public function getParent(): ?UnalterableMigrationInterface;
 
     public function getUpSql(): string;
