+++ packages/Doctrine/UnalterableMigrations/src/UnalterableMigrationTrait.php
@@ -4,52 +4,14 @@ namespace PetrKnap\Doctrine\UnalterableMigrations;
 
 use Doctrine\DBAL\Schema\Schema;
 
-trait UnalterableSqlMigrationTrait
+trait UnalterableMigrationTrait
 {
-    public function getParentClassName(): ?string
-    {
-        return null;
-    }
-
-    public function getParent(): ?UnalterableSqlMigrationInterface
-    {
-        $parentClassName = $this->getParentClassName();
-
-        if ($parentClassName) {
-            return new $parentClassName(clone $this->version);
-        }
-
-        return null;
-    }
-
-    public function getDownSql(): ?string
-    {
-        return null;
-    }
-
     public function up(Schema $schema): void
     {
         $this->parentsDown();
 
-        if ($this->getUpSql() !== UnalterableSqlMigrationInterface::REMOVE_PARENT) {
+        if ($this->getUpSql() !== UnalterableMigrationInterface::REMOVE_PARENT) {
             $this->addSql($this->getUpSql());
-        }
-    }
-
-    public function down(Schema $schema): void
-    {
-        if ($this->getDownSql()) {
-            $this->addSql($this->getDownSql());
-        } else {
-            $this->abortIf($this->getParent() === null, 'Can not generate down method without down SQL');
-        }
-
-        if ($this->getUpSql() !== UnalterableSqlMigrationInterface::REMOVE_PARENT) {
-            $this->parentsDown();
-        }
-
-        if ($this->getParent()) {
-            $this->addSql($this->getParent()->getUpSql());
         }
     }
 
@@ -62,5 +24,43 @@ trait UnalterableSqlMigrationTrait
             }
             $parent = $parent->getParent();
         }
+    }
+
+    public function getParent(): ?UnalterableMigrationInterface
+    {
+        $parentClassName = $this->getParentClassName();
+
+        if ($parentClassName) {
+            return new $parentClassName(clone $this->version);
+        }
+
+        return null;
+    }
+
+    public function getParentClassName(): ?string
+    {
+        return null;
+    }
+
+    public function down(Schema $schema): void
+    {
+        if ($this->getDownSql()) {
+            $this->addSql($this->getDownSql());
+        } else {
+            $this->abortIf($this->getParent() === null, 'Can not generate down method without down SQL');
+        }
+
+        if ($this->getUpSql() !== UnalterableMigrationInterface::REMOVE_PARENT) {
+            $this->parentsDown();
+        }
+
+        if ($this->getParent()) {
+            $this->addSql($this->getParent()->getUpSql());
+        }
+    }
+
+    public function getDownSql(): ?string
+    {
+        return null;
     }
 }
